Extract setupLSPatch task to separate script
This change moves the `setupLSPatch` task logic and its imports from `app/build.gradle.kts` into a new script `scripts/setup_lspatch.gradle.kts`. This improves the maintainability and readability of the main build file by isolating this complex logic.
Verified by running `./gradlew tasks` to ensure the task is still registered and valid.
